+++ android/src/main/kotlin/com/gdelataillade/alarm/alarm/AlarmPlugin.kt
@@ -1,31 +1,48 @@
 package com.gdelataillade.alarm.alarm
 
+import android.content.Context
+import android.content.Intent
 import androidx.annotation.NonNull
-
+import io.flutter.Log
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
 
-/** AlarmPlugin */
+/// Communication between Flutter Alarm service and native Android.
 class AlarmPlugin: FlutterPlugin, MethodCallHandler {
-  /// The MethodChannel that will the communication between Flutter and native Android
-  ///
-  /// This local reference serves to register the plugin with the Flutter Engine and unregister it
-  /// when the Flutter Engine is detached from the Activity
+  private lateinit var context: Context
   private lateinit var channel : MethodChannel
 
   override fun onAttachedToEngine(@NonNull flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
-    channel = MethodChannel(flutterPluginBinding.binaryMessenger, "alarm")
-    channel.setMethodCallHandler(this)
+      context = flutterPluginBinding.applicationContext
+      channel = MethodChannel(flutterPluginBinding.binaryMessenger, "com.gdelataillade.alarm/notifOnAppKill")
+      channel.setMethodCallHandler(this)
   }
 
   override fun onMethodCall(@NonNull call: MethodCall, @NonNull result: Result) {
-    if (call.method == "getPlatformVersion") {
-      result.success("Android ${android.os.Build.VERSION.RELEASE}")
-    } else {
-      result.notImplemented()
+    when (call.method) {
+        "setNotificationOnKillService" -> {
+          val title = call.argument<String>("title")
+          val description = call.argument<String>("description")
+
+          val serviceIntent = Intent(context, NotificationOnKillService::class.java)
+
+          serviceIntent.putExtra("title", title)
+          serviceIntent.putExtra("description", description)
+
+          context.startService(serviceIntent)
+          result.success(true)
+        }
+        "stopNotificationOnKillService" -> {
+          val serviceIntent = Intent(context, NotificationOnKillService::class.java)
+          context.stopService(serviceIntent)
+          result.success(true)
+        }
+        else -> {
+            result.notImplemented()
+        }
     }
   }