+++ android/src/main/kotlin/com/gdelataillade/alarm/alarm/NotificationOnKillService.kt
@@ -1,0 +1,65 @@
+package com.gdelataillade.alarm.alarm
+
+import android.annotation.SuppressLint
+import android.app.NotificationChannel
+import android.app.NotificationManager
+import android.app.PendingIntent
+import android.app.Service
+import android.content.Context
+import android.content.Intent
+import android.os.Build
+import android.os.IBinder
+import androidx.annotation.RequiresApi
+import androidx.core.app.NotificationCompat
+import io.flutter.Log
+
+class NotificationOnKillService: Service() {
+
+    private lateinit var title: String
+    private lateinit var description: String
+
+    override fun onStartCommand(intent: Intent?, flags: Int, startId: Int): Int {
+        title = intent?.getStringExtra("title") ?: "The app was killed"
+        description = intent?.getStringExtra("description") ?: "Your alarm may not ring, you should reopen the app."
+
+        return START_STICKY
+    }
+
+    @SuppressLint("UnspecifiedImmutableFlag")
+    @RequiresApi(Build.VERSION_CODES.O)
+    override fun onTaskRemoved(rootIntent: Intent?) {
+        try {
+
+            val notificationIntent = packageManager.getLaunchIntentForPackage(packageName)
+            val pendingIntent = PendingIntent.getActivity(this, 0, notificationIntent, 0)
+
+            val notificationBuilder = NotificationCompat.Builder(this, "com.gdelataillade.alarm.alarm")
+                .setSmallIcon(android.R.drawable.ic_notification_overlay)
+                .setContentTitle(title)
+                .setContentText(description)
+                .setAutoCancel(false)
+                .setPriority(NotificationCompat.PRIORITY_DEFAULT)
+                .setContentIntent(pendingIntent)
+
+            val name = "Alarm notification service on application kill"
+            val descriptionText = "If an alarm was set and the app is killed, a notification will show to warn the user the alarm will not ring as long as the app is killed"
+            val importance = NotificationManager.IMPORTANCE_DEFAULT
+            val channel = NotificationChannel("com.gdelataillade.alarm.alarm", name, importance).apply {
+                description = descriptionText
+            }
+
+            // Register the channel with the system
+            val notificationManager: NotificationManager =
+                getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
+            notificationManager.createNotificationChannel(channel)
+            notificationManager.notify(123, notificationBuilder.build())
+        } catch (e: Exception) {
+            Log.d("NotificationOnKillService", "Error showing notification", e)
+        }
+        super.onTaskRemoved(rootIntent)
+    }
+
+    override fun onBind(intent: Intent?): IBinder? {
+        return null
+    }
+}