+++ example/android/app/src/main/kotlin/com/gdelataillade/alarm/alarm_example/MainActivity.kt
@@ -2,5 +2,11 @@ package com.gdelataillade.alarm.alarm_example
 
 import io.flutter.embedding.android.FlutterActivity
 
+import android.content.Intent
+import android.os.Bundle
+
 class MainActivity: FlutterActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+    }
 }