+++ ios/Classes/SwiftAlarmPlugin.swift
@@ -10,6 +10,9 @@ public class SwiftAlarmPlugin: NSObject, FlutterPlugin {
   }
 
   public var audioPlayer: AVAudioPlayer!
+  public var notifOnKillEnabled: Bool!
+  public var notificationTitleOnKill: String!
+  public var notificationBodyOnKill: String!
 
   private func setUpAudio() {
     try! AVAudioSession.sharedInstance().setCategory(.playback, options: [.mixWithOthers])
@@ -21,12 +24,9 @@ public class SwiftAlarmPlugin: NSObject, FlutterPlugin {
       if call.method == "setAlarm" {
         self.setAlarm(call: call, result: result)
       } else if call.method == "stopAlarm" {
-        if self.audioPlayer != nil {
-          self.audioPlayer.stop()
-          self.audioPlayer = nil
-          result(true)
-        }
-        result(false)
+        self.audioPlayer.stop()
+        NotificationCenter.default.removeObserver(self, name: UIApplication.willTerminateNotification, object: nil)
+        result(true)
       } else if call.method == "audioCurrentTime" {
         if self.audioPlayer != nil {
           result(Double(self.audioPlayer.currentTime))
@@ -45,6 +45,15 @@ public class SwiftAlarmPlugin: NSObject, FlutterPlugin {
     self.setUpAudio()
 
     let args = call.arguments as! Dictionary<String, Any>
+
+    notifOnKillEnabled = args["notifOnKillEnabled"] as! Bool
+    notificationTitleOnKill = args["notifTitleOnAppKill"] as! String
+    notificationBodyOnKill = args["notifDescriptionOnAppKill"] as! String
+
+    if notifOnKillEnabled {
+      NotificationCenter.default.addObserver(self, selector: #selector(applicationWillTerminate(_:)), name: UIApplication.willTerminateNotification, object: nil)
+    }
+
     let assetAudio = args["assetAudio"] as! String
     let delayInSeconds = args["delayInSeconds"] as! Double
     let loopAudio = args["loopAudio"] as! Bool
@@ -71,5 +80,14 @@ public class SwiftAlarmPlugin: NSObject, FlutterPlugin {
     self.audioPlayer.play(atTime: time)
 
     result(true)
+  }
+
+  @objc func applicationWillTerminate(_ notification: Notification) {
+    let content = UNMutableNotificationContent()
+    content.title = notificationTitleOnKill
+    content.body = notificationBodyOnKill
+    let trigger = UNTimeIntervalNotificationTrigger(timeInterval: 1, repeats: false)
+    let request = UNNotificationRequest(identifier: "notification on app kill", content: content, trigger: trigger)
+    UNUserNotificationCenter.current().add(request)
   }
 }